fixed double  issue

+++ FVCalc.java
@@ -1,10 +1,10 @@
 // Computes the future value of a saving investment.
 public class FVCalc {
 	public static void main(String[] args){
-		double currentValue =Double.parseDouble(args[0]);
+		int currentValue =Integer.parseInt(args[0]);
 		double rate =Double.parseDouble(args[1]);
-		double years =Double.parseDouble(args[2]);
-		double futureValue = currentValue* Math.pow(1+rate/100, years);
+		int years =Integer.parseInt(args[2]);
+		int futureValue = (int) (currentValue* Math.pow(1+rate/100, years));
 		System.out.println("After "+ years +" years, "+"$"+currentValue+"saved at  "+rate+"% "+"will yield "+"$"+(int)futureValue);
 	}
 }
